+++ capi/autd3capi-def/build.rs
@@ -4,21 +4,24 @@
  * Created Date: 29/05/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 26/10/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-use std::env;
-
-use wrapper_generator::generate;
-
+#[cfg(feature = "generate_wrapper")]
 fn main() {
+    use autd3capi_wrapper_generator::generate;
+    use std::env;
+
     let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
 
     if let Err(e) = generate(crate_dir) {
         eprintln!("{}", e);
     }
 }
+
+#[cfg(not(feature = "generate_wrapper"))]
+fn main() {}