+++ capi/autd3capi/src/modulation/custom.rs
@@ -4,7 +4,7 @@
  * Created Date: 23/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 23/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -13,7 +13,7 @@
 
 #![allow(clippy::missing_safety_doc)]
 
-use autd3capi_def::{common::CustomModulation, ModulationPtr, SamplingConfiguration};
+use autd3capi_def::{CustomModulation, ModulationPtr, SamplingConfiguration};
 
 #[no_mangle]
 #[must_use]
@@ -23,8 +23,7 @@ pub unsafe extern "C" fn AUTDModulationCustom(
     ptr: *const u8,
     len: u64,
 ) -> ModulationPtr {
-    let mut buf =
-        Vec::<autd3capi_def::common::driver::common::EmitIntensity>::with_capacity(len as _);
+    let mut buf = Vec::<autd3capi_def::driver::common::EmitIntensity>::with_capacity(len as _);
     buf.set_len(len as _);
     std::ptr::copy_nonoverlapping(ptr as _, buf.as_mut_ptr(), len as _);
     ModulationPtr::new(CustomModulation {