+++ capi/autd3capi/src/modulation/fourier.rs
@@ -4,7 +4,7 @@
  * Created Date: 23/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 22/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -14,11 +14,8 @@
 #![allow(clippy::missing_safety_doc)]
 
 use autd3capi_def::{
-    common::{
-        autd3::modulation::{Fourier, Sine},
-        *,
-    },
-    take_mod, ModulationPtr,
+    autd3::modulation::{Fourier, Sine},
+    *,
 };
 
 #[no_mangle]