+++ dotnet/cs/src/Gain/Holo/Naive.cs
@@ -4,7 +4,7 @@
  * Created Date: 13/09/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 07/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -14,6 +14,8 @@
 #if UNITY_2020_2_OR_NEWER
 #nullable enable
 #endif
+
+using AUTD3Sharp.NativeMethods;
 
 namespace AUTD3Sharp.Gain.Holo
 {