+++ python/pyautd3/native_methods/autd3capi_link_simulator.py
@@ -2,7 +2,7 @@
 import threading
 import ctypes
 import os
-from .autd3capi_def import GeometryPtr, LinkBuilderPtr, LinkPtr, ResultI32
+from pyautd3.native_methods.autd3capi_def import GeometryPtr, LinkBuilderPtr, LinkPtr, ResultI32
 
 
 class LinkSimulatorBuilderPtr(ctypes.Structure):