+++ python/pyautd3/native_methods/autd3capi_link_soem.py
@@ -2,7 +2,7 @@
 import threading
 import ctypes
 import os
-from .autd3capi_def import LinkBuilderPtr, TimerStrategy
+from pyautd3.native_methods.autd3capi_def import LinkBuilderPtr, TimerStrategy
 
 from enum import IntEnum